•	To create a system that provide the common key to each individual. •	The dynamic nature of system allows the existing member to leave the
As a result of the increased popularity of group-oriented applications
and protocols, group communication occurs in many different settings:
from network layer multicasting to application layer tele- and
videoconferencing. There is a need for security services to provide
group-oriented communication privacy and data integrity. To provide this
form of group communication privacy, it is important that members of the
group can establish a common secret key for encrypting group
communication data. For example, consider a group of people in a
peer-to-peer ad hoc network having a closed and confidential business
meeting. Since they have not previously agreed upon a common secret key,
communication between group members is susceptible to eavesdropping. To
solve the problem, we need a secure distributed group key agreement
protocol such that the group of people can establish the common group
key for secure and private communication. Note that this type of key
agreement protocols is both distributed and contributory in nature: each
member of the group contributes its part to the overall group key.
Distributed group key agreement protocols is different from Centralized
group key distribution protocols as  Centralized protocols rely on a
centralized key server to efficiently distribute the group key. For
distributed key agreement protocols, however, no centralized key server
is available. This arrangement is justified in many situations – e.g.,
in a peer-to-peer or ad hoc network where centralized resources are not
readily available. Moreover, an advantage of distributed protocols over
the centralized protocols is the increase in system reliability, since
the group key is generated in a shared and contributory fashion and
there is no single point of failure.